+++ tests/scm/config/security/test_vulnerability_protection_profiles.py
@@ -17,7 +17,7 @@ from scm.models.security.vulnerability_protection_profiles import (
     VulnerabilityProfileResponseModel,
     VulnerabilityProfileSeverity,
 )
-from tests.factories import (
+from tests.test_factories.security import (
     VulnerabilityProfileCreateApiFactory,
     VulnerabilityProfileResponseFactory,
     VulnerabilityProfileRuleModelFactory,
