+++ tests/scm/models/security/test_vulnerability_protection_profile_models.py
@@ -18,7 +18,7 @@ from scm.models.security.vulnerability_protection_profiles import (
     VulnerabilityProfileSeverity,
     VulnerabilityProfileUpdateModel,
 )
-from tests.factories import (
+from tests.test_factories.security import (
     VulnerabilityProfileCreateModelFactory,
     VulnerabilityProfileUpdateModelFactory,
 )
